Update the `give()` function to be lowercase.
In order for PHPStorm to autocomplete the dynamic resolution of the container,
the override function name case must match the implementation. Since user
defined function names in PHP are case-insensitive, there is no backwards
compatibility issue with updating the function name `Give()` to match the
convention of a lowercase function name.