refactor: remove tabindex from legacy fields
These elements are already focusable, so `tabindex` with the values of
`0` and `-1` are unnecessary. These were using `1` which increases the
tab order on the page. This might have been to overcome some issue of
tab ordering within the markup of the form. That’s a dangerous game
though, since the form is but a small part of the entire page and we
don’t have control over the tab index of elements outside of the form.
Because of this, I think it’s best to not tamper with the tabindex. We
can allow of a degraded tab experience using the form and preserve the
experience for the rest of the document.

See: https://www.scottohara.me/blog/2019/05/25/tabindex.html