Updated currency symbol for LKR (#6096)
LKR has three currency symbols, all three are accepted. The plugin has been using the currency symbol which is used only used by a part of Sri Lanka. I've now changed to Rs which can be understood and accepted by anyone instead of just a specific of people in the country.